requirements.txt
python-slugify
tmdbsimple
tqdm
